Cleanup: move files to machine specific directories
To use the emacs config, export it to an .el file using

   (org-babel-load-file)

Then, copy over the file to the machine specific directory (e. g. 'workstation')
and run

   nix build .#homeManagerConfigurations.USERNAME.activationPackage
   ./result/activate

to activate it.
This will allow for different machines to have specific emacs configs that don't clash
with each other (it's a hacky 'fix' though, you'll have to edit settings.el, but I
don't have time for something more sophisticated rn).